Move flashable inside vecengine package
