+++ emitter/ancestor/quorum_indexer_test.go
@@ -13,10 +13,10 @@ import (
 	"github.com/Fantom-foundation/lachesis-base/inter/dag/tdag"
 	"github.com/Fantom-foundation/lachesis-base/inter/idx"
 	"github.com/Fantom-foundation/lachesis-base/inter/pos"
-	"github.com/Fantom-foundation/lachesis-base/kvdb/flashable"
 	"github.com/Fantom-foundation/lachesis-base/kvdb/memorydb"
 	"github.com/Fantom-foundation/lachesis-base/utils"
 	"github.com/Fantom-foundation/lachesis-base/utils/adapters"
+	"github.com/Fantom-foundation/lachesis-base/vecengine/flashable"
 	"github.com/Fantom-foundation/lachesis-base/vecfc"
 )
