+++ vecengine/flashable/flashable.go
