+++ vecengine/flashable/flashable_test.go
